moves auth to new setup hook
