+++ src/hooks/app/useSetUp.ts
@@ -4,34 +4,16 @@ import { listen } from '@tauri-apps/api/event';
 import { TauriEvent } from '../../types.ts';
 
 import { setSetupComplete, setSetupDetails, useAppStateStore } from '../../store/appStateStore.ts';
-import { fetchBackendInMemoryConfig, getExistingTokens } from '@app/store/useAirdropStore.ts';
-import { handleRefreshAirdropTokens } from '@app/hooks/airdrop/stateHelpers/useAirdropTokensRefresh.ts';
 
 export function useSetUp() {
     const isInitializingRef = useRef(false);
     const adminShow = useUIStore((s) => s.adminShow);
     const fetchApplicationsVersionsWithRetry = useAppStateStore((s) => s.fetchApplicationsVersionsWithRetry);
-    const mmProxyVersion = useAppStateStore((s) => s.applications_versions?.mm_proxy);
 
     const handlePostSetup = useCallback(async () => {
         await setSetupComplete();
         await fetchApplicationsVersionsWithRetry();
     }, [fetchApplicationsVersionsWithRetry]);
-
-    useEffect(() => {
-        async function initWithToken() {
-            const beConfig = await fetchBackendInMemoryConfig();
-            await getExistingTokens();
-            if (beConfig?.airdropUrl) {
-                await handleRefreshAirdropTokens(beConfig.airdropUrl);
-            }
-        }
-        if (mmProxyVersion) {
-            // Timeout added as there are instances where this gets called more than once
-            const timeout = setTimeout(() => initWithToken(), 500);
-            return () => clearTimeout(timeout);
-        }
-    }, [mmProxyVersion]);
 
     useEffect(() => {
         if (adminShow === 'setup') return;
